+++ crates/cli/tests/query_test.rs
@@ -313,7 +313,7 @@ mod projects {
         let ids: Vec<String> = json.projects.iter().map(|p| p.id.to_string()).collect();
 
         assert_eq!(ids, string_vec!["advanced", "noConfig"]);
-        assert!(json.options.affected);
+        assert!(json.options.affected.is_some());
     }
 
     #[test]
@@ -347,7 +347,7 @@ mod projects {
 
         assert_eq!(
             assert.output(),
-            "advanced | advanced | frontend | application | typescript\nnoConfig | no-config | unknown | unknown | unknown\n\n"
+            "advanced | advanced | frontend | application | typescript | ...\nnoConfig | no-config | unknown | unknown | unknown | ...\n\n"
         );
     }
 
@@ -377,8 +377,7 @@ mod projects {
         let ids: Vec<String> = json.projects.iter().map(|p| p.id.to_string()).collect();
 
         assert_eq!(ids, string_vec!["advanced", "basic", "noConfig"]);
-        assert!(json.options.affected);
-        assert!(json.options.dependents);
+        assert!(json.options.affected.is_some());
     }
 
     #[test]
@@ -415,7 +414,7 @@ mod projects {
         let ids: Vec<String> = json.projects.iter().map(|p| p.id.to_string()).collect();
 
         assert_eq!(ids, string_vec!["advanced", "noConfig"]);
-        assert!(json.options.affected);
+        assert!(json.options.affected.is_some());
     }
 
     #[test]
@@ -660,8 +659,8 @@ mod projects {
             let json: QueryProjectsResult = json::parse(assert.output()).unwrap();
             let ids: Vec<String> = json.projects.iter().map(|p| p.id.to_string()).collect();
 
-            assert_eq!(ids, string_vec!["noConfig"]);
-            assert!(json.options.affected);
+            assert_eq!(ids, string_vec!["advanced", "noConfig"]);
+            assert!(json.options.affected.is_some());
         }
     }
 }