+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## Unreleased
+
+#### 🚀 Updates
+
+- Implemented a new affected project tracker that is more accurate and more powerful.
+  - Can now control the depth of upstream (dependencies) and downstream (dependents).
+  - Affected information now logs more information and is included in reports/JSON.
+  - Added `--upstream` and `--downstream` options to `moon query projects`.
+  - Coming soon for affected tasks as well!
+- Updated `moon query projects` to include the project description as a trailing value.
+- Updated `moon query tasks` to include the task type and platform, and the task description as a
+  trailing value.
+
 ## 1.28.3
 
 #### 🐞 Fixes